requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
gradio>=3.40.0
streamlit>=1.25.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
accelerate>=0.20.0
librosa
sqlite3